Add routing, Home and Login views
