+++ src/components/Login.js
@@ -1,0 +1,4 @@
+
+export function Login(){
+    return <h1>hola login</h1>
+}
